requirements.txt
# AI Tool TrendAnalyzer - Dependencies
# Install: pip install -r requirements.txt

# Web Scraping (Dynamic pages)
playwright>=1.40.0
beautifulsoup4==4.12.3       # HTML parser
httpx>=0.27.0
lxml==5.1.0                  # Fast XML/HTML parser (BS4 backend)

# GenAI (Groq - Free LLM API)
groq>=0.4.0

# API Layer
fastapi>=0.109.0
uvicorn[standard]>=0.27.0

# Dashboard
streamlit>=1.31.0
plotly>=5.18.0
pandas>=2.1.0

# Scheduling
apscheduler>=3.10.0

# Utilities
python-dotenv>=1.0.0
pydantic==2.6.4        # Data validation (used by FastAPI)